Refactor Clifford operations to separate out those that deal with pulses from pure matrix manipulation
